requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
Flask-Caching==2.1.0
solana==0.30.2
base58==2.1.1
gunicorn==21.2.0 